Extract the shared product membership card
The category and collection panels were the same component twice — 797 lines
holding two copies of the drag-reorder rollback, the visible-rows-only select
all, the pagination offset and the dnd-kit row workaround. A bug in any of them
had to be fixed in both places, and the pagination work had just widened the
duplication.

One card now takes the five membership hooks, a locale namespace and a readOnly
flag; both call sites are thin adapters. `readOnly` replaces the
collection-specific `automatic` so the card stays agnostic about why a
grouping's membership is derived.
